Fix the implementation of the 'suggestGranularity' warning message to be more actionable.
Currently, the warning is issued if any single measurement is > 1000xtimer
granularity (typically 50-70 microseconds) during the warmup phase.  During
warmup there are a lot of reasons this might happen even if normal timing
phase measurements are typically less than the limit (in fact this might even
be expected).  Instead we change the warning to only fire if all 'timing'
measurements are > 1000x timer granularity.  This should ensure that the
warning is only issued for benchmarks that are actually large.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=128992528